# [IMP] pos_self_order: Avoid extra _get_attributes_by_ptal_id calls