--- a/addons/pos_self_order_sale/models/product_product.py
+++ b/addons/pos_self_order_sale/models/product_product.py
@@ -6,8 +6,8 @@ from odoo import models
 class ProductProduct(models.Model):
     _inherit = "product.product"
 
-    def _get_product_for_ui(self, pos_config):
+    def _get_product_for_ui(self, pos_config, attributes=None):
         self.ensure_one()
-        product = super()._get_product_for_ui(pos_config)
+        product = super()._get_product_for_ui(pos_config, attributes)
         product["optional_product_ids"] = self.optional_product_ids.product_variant_ids.ids
         return product