@zaha-odoo 2024-03-22
Closing due to possibility of backport to 16.4, will open a new PR to target that instead.